Refactor to clarify
