+++ jest.config.js
@@ -8,7 +8,13 @@ module.exports = {
   },
   testMatch: ["<rootDir>/src/**/*.test.{js,ts}"],
   setupFiles: ["dotenv/config"],
-  collectCoverageFrom: ["src/**/*.ts", "!src/**/index.ts"],
+  collectCoverageFrom: [
+    "src/**/*.ts",
+    "!src/ProviderGraphObjectCache.ts",
+    "!src/getIterationState.ts",
+    "!src/**/index.ts",
+    "!src/steps/**/*.ts",
+  ],
   moduleFileExtensions: [...defaults.moduleFileExtensions, "ts"],
   testEnvironment: "node",
   clearMocks: true,
